More Admin
So in this one ive tested editing and deleting a product, as well as editing a users details. At the moment there is nothing other than view details within a bootstrap modal. Plan to add a dropdown to change job role, only allow the owner to access this page & the final touches on the rest of the admin pages. Will need CSS ASAP as it looks rubbish.